tg/requirements.txt
aiogram>=3.4.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
PyJWT>=2.8.0
bcrypt>=4.1.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
